refactor: refactor constants, add measures
- remove unused constants
- collect gloabl constants to single file
- add measurements to coin, price
