--- src/App.constants.js
@@ -1,7 +1,0 @@
-export const PAGE_TITLE = '🥤자판기🥤';
-
-export const ID_PRODUCT_ADD_TAB = 'product-add-section';
-export const ID_MACHINE_TAB = 'vending-machine-manage-section';
-export const ID_PURCHASE_TAB = 'product-purchase-section';
-
-export const STRING_SECTION_SUFFIX = '-section';
