+++ src/ChargeMachine/components.js
@@ -1,5 +1,15 @@
 import { customCreateElement } from '../CreateElementUtils.js';
 import {
+  STRING_COIN_10,
+  STRING_COIN_100,
+  STRING_COIN_50,
+  STRING_COIN_500,
+  STYLE_TABLE,
+  STYLE_TABLE_BORDER,
+  VAL_COLUMN_SIZE,
+  VAL_ROW_SIZE,
+} from '../globalConstants.js';
+import {
   ID_MACHINE_CHARGE_AMOUNT,
   ID_MACHINE_CHARGE_INPUT,
   ID_MACHINE_CHARGE_SUBMIT,
@@ -10,22 +20,14 @@ import {
   ID_MACHINE_COIN_STATUS,
   PLACEHOLDER_INSERT_COIN,
   STRING_CHARGE_SUM_LABEL,
-  STRING_COIN_10,
-  STRING_COIN_100,
-  STRING_COIN_50,
-  STRING_COIN_500,
   STRING_COIN_NAMES_HEADER,
   STRING_COIN_QUANTITY_HEADER,
-  STYLE_TABLE,
-  STYLE_TABLE_BORDER,
   TITLE_COIN_IN_MACHINE,
   TITLE_FORM,
-  VAL_COLUMN_SIZE,
   VAL_INSERT_BUTTON,
-  VAL_ROW_SIZE,
 } from './constants.js';
 
-// ------------- element generators -------------
+// ------------- form components --------------
 const createData = function createDataArrayForTable(value, id) {
   return [
     { value, width: `${VAL_COLUMN_SIZE}px` },
@@ -33,6 +35,45 @@ const createData = function createDataArrayForTable(value, id) {
   ];
 };
 
+export const ChargeFormSection = function createNewSection() {
+  return document.createElement('div');
+};
+
+export const ChargeFormTitle = customCreateElement({
+  tag: 'h2',
+  value: TITLE_FORM,
+});
+
+export const ChargeInput = customCreateElement({
+  tag: 'input',
+  attributes: {
+    id: ID_MACHINE_CHARGE_INPUT,
+    type: 'text',
+    placeholder: PLACEHOLDER_INSERT_COIN,
+  },
+});
+
+export const ChargeFormSubmit = function createChargeSubmit() {
+  return customCreateElement({
+    tag: 'button',
+    attributes: {
+      id: ID_MACHINE_CHARGE_SUBMIT,
+    },
+    value: VAL_INSERT_BUTTON,
+  });
+};
+
+export const ChargeSumLabel = customCreateElement({
+  tag: 'p',
+  value: STRING_CHARGE_SUM_LABEL,
+});
+
+export const ChargeSumSpan = customCreateElement({
+  tag: 'span',
+  attributes: { id: ID_MACHINE_CHARGE_AMOUNT },
+});
+
+// -------------- table components --------------
 const createTD = function createSingleTableDataElement(data) {
   return customCreateElement({
     tag: 'td',
@@ -72,47 +113,6 @@ const createTRs = function createTableRows(dataArray) {
   return dataArray.map((data) => createTR(data));
 };
 
-// ----------------- components -----------------
-// form components
-export const ChargeFormSection = function createNewSection() {
-  return document.createElement('div');
-};
-
-export const ChargeFormTitle = customCreateElement({
-  tag: 'h2',
-  value: TITLE_FORM,
-});
-
-export const ChargeInput = customCreateElement({
-  tag: 'input',
-  attributes: {
-    id: ID_MACHINE_CHARGE_INPUT,
-    type: 'text',
-    placeholder: PLACEHOLDER_INSERT_COIN,
-  },
-});
-
-export const ChargeFormSubmit = function createChargeSubmit() {
-  customCreateElement({
-    tag: 'button',
-    attributes: {
-      id: ID_MACHINE_CHARGE_SUBMIT,
-    },
-    value: VAL_INSERT_BUTTON,
-  });
-};
-
-export const ChargeSumLabel = customCreateElement({
-  tag: 'p',
-  value: STRING_CHARGE_SUM_LABEL,
-});
-
-export const ChargeSumSpan = customCreateElement({
-  tag: 'span',
-  attributes: { id: ID_MACHINE_CHARGE_AMOUNT },
-});
-
-// table components
 export const CoinTableSection = document.createElement('div');
 
 export const CoinTableTitle = customCreateElement({
