+++ src/globalConstants.js
@@ -1,0 +1,19 @@
+export const PAGE_TITLE = '🥤자판기🥤';
+
+export const ID_PRODUCT_ADD_TAB = 'product-add-section';
+export const ID_MACHINE_TAB = 'vending-machine-manage-section';
+export const ID_PURCHASE_TAB = 'product-purchase-section';
+
+export const STRING_SECTION_SUFFIX = '-section';
+
+export const VAL_PRICE_ROUND_STANDARD = 10;
+
+export const STRING_COIN_500 = '500';
+export const STRING_COIN_100 = '100';
+export const STRING_COIN_50 = '50';
+export const STRING_COIN_10 = '10';
+
+export const STYLE_TABLE = 'border-collapse: collapse; text-align: center';
+export const STYLE_TABLE_BORDER = 'border: 1px solid black;';
+export const VAL_COLUMN_SIZE = '100';
+export const VAL_ROW_SIZE = '50';
